+++ src/columns.ts
@@ -107,10 +107,15 @@ export abstract class Column implements IStatementGiver {
 }
 
 export class BooleanColumn extends Column implements Condition {
-  // implement Condition
+  // START implement Condition
   public readonly leftExpression: Expression = new Expression(this)
   public readonly leftOperand: Operand = this.leftExpression.leftOperand
   public readonly type: ExpressionType = ExpressionType.BOOLEAN
+
+  public getColumns(): BooleanColumn[] {
+    return [this]
+  }
+  // END implement Condition
 
   constructor(data: ColumnObj) {
     super(data)
@@ -140,10 +145,6 @@ export class BooleanColumn extends Column implements Condition {
 
   public not(): Condition {
     return new Condition(new Expression(this, true))
-  }
-
-  public getColumns(): Column[] {
-    return [this]
   }
 }